# src/docs/swagger.yaml
definitions:
...
  model.CommonPrefix:
    description: Common prefix representing a virtual folder
    properties:
      prefix:
        description: The prefix (folder path)
        example: documents/
        type: string
    type: object
...
  model.ListBucketResult:
    description: Success response for ListObjects operation (S3 ListObjectsV2 compatible)
    properties:
      commonPrefixes:
        description: Common prefixes (virtual folders) when using delimiter
        items:
          $ref: '#/definitions/model.CommonPrefix'
        type: array
      contents:
        description: List of objects
        items:
          $ref: '#/definitions/model.ObjectContent'
        type: array
      delimiter:
        description: Delimiter used to group common prefixes
        example: /
        type: string
      isTruncated:
        description: Indicates if the response is truncated
        example: false
        type: boolean
      keyCount:
        description: Number of keys returned
        example: 10
        type: integer
      maxKeys:
        description: Maximum number of keys returned
        example: 1000
        type: integer
      name:
        description: Name of the bucket
        example: my-bucket
        type: string
      prefix:
        description: Prefix used to filter objects
        example: documents/
        type: string
    type: object
  model.ObjectContent:
    description: Individual object information in list response
    properties:
      etag:
        description: ETag (hash) of the object
        example: '"d41d8cd98f00b204e9800998ecf8427e"'
        type: string
      key:
        description: Object key (path)
        example: documents/file.txt
        type: string
      lastModified:
        description: Last modification time in RFC3339 format
        example: "2024-01-15T10:30:00Z"
        type: string
      size:
        description: Size in bytes
        example: 1024
        type: integer
      storageClass:
        description: Storage class
        example: STANDARD
        type: string
    type: object
...
  model.S3Error:
    description: S3-compatible error response returned on failure
    properties:
      code:
        description: Error code identifying the type of error
        example: NoSuchBucket
        type: string
      message:
        description: Human-readable error message
        example: The specified bucket does not exist
        type: string
      requestId:
        description: Unique request identifier for troubleshooting
        example: 4442587FB7D0A2F9
        type: string
      resource:
        description: The resource associated with the error
        example: my-bucket
        type: string
    type: object
info:
  contact:
    name: API Support
  termsOfService: http://swagger.io/terms/
paths:
...
  /{bucket}:
    delete:
      description: Deletes the specified bucket. The bucket must be empty before it
        can be deleted
      operationId: deletebucket
      parameters:
      - description: Name of the bucket to delete
        in: path
        name: bucket
        required: true
        type: string
      - description: AWS4-HMAC-SHA256 authorization header (required if S3_AUTH_ENABLED=true)
        in: header
        name: Authorization
        type: string
      - description: Request timestamp in ISO 8601 format (required if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Date
        type: string
      - description: SHA256 hash of request payload or 'UNSIGNED-PAYLOAD' (required
          if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Content-SHA256
        type: string
      produces:
      - text/xml
      responses:
        "204":
          description: Bucket deleted successfully
        "403":
          description: Access denied or signature mismatch
          schema:
            $ref: '#/definitions/model.S3Error'
        "404":
          description: Bucket not found
          schema:
            $ref: '#/definitions/model.S3Error'
        "409":
          description: Bucket not empty
          schema:
            $ref: '#/definitions/model.S3Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.S3Error'
      summary: Delete a bucket
      tags:
      - buckets
    get:
      description: Returns a list of objects in the specified bucket (S3 ListObjectsV2
        compatible)
      operationId: listobjects
      parameters:
      - description: Bucket name
        in: path
        name: bucket
        required: true
        type: string
      - description: S3 list API version (must be 2 for ListObjectsV2)
        enum:
        - 2
        in: query
        name: list-type
        type: integer
      - description: Limits the response to keys that begin with the specified prefix
        in: query
        name: prefix
        type: string
      - description: Character used to group keys (usually /)
        in: query
        name: delimiter
        type: string
      - default: 1000
        description: Maximum number of keys returned
        in: query
        name: max-keys
        type: integer
      - description: AWS4-HMAC-SHA256 authorization header (required if S3_AUTH_ENABLED=true)
        in: header
        name: Authorization
        type: string
      - description: Request timestamp in ISO 8601 format (required if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Date
        type: string
      produces:
      - text/xml
      responses:
        "200":
          description: List of objects
          schema:
            $ref: '#/definitions/model.ListBucketResult'
        "403":
          description: Access denied or signature mismatch
          schema:
            $ref: '#/definitions/model.S3Error'
        "404":
          description: Bucket not found
          schema:
            $ref: '#/definitions/model.S3Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.S3Error'
      summary: List objects in a bucket
      tags:
      - objects
    head:
      description: Checks if a bucket exists and you have permission to access it
      operationId: headbucket
      parameters:
      - description: Name of the bucket to check
        in: path
        name: bucket
        required: true
        type: string
      - description: AWS4-HMAC-SHA256 authorization header (required if S3_AUTH_ENABLED=true)
        in: header
        name: Authorization
        type: string
      - description: Request timestamp in ISO 8601 format (required if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Date
        type: string
      - description: SHA256 hash of request payload or 'UNSIGNED-PAYLOAD' (required
          if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Content-SHA256
        type: string
      responses:
        "200":
          description: Bucket exists
        "403":
          description: Access denied or signature mismatch
          schema:
            $ref: '#/definitions/model.S3Error'
        "404":
          description: Bucket not found
          schema:
            $ref: '#/definitions/model.S3Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.S3Error'
      summary: Check if bucket exists
      tags:
      - buckets
    put:
      description: Creates a new S3 bucket with the specified name
      operationId: createbucket
      parameters:
      - description: Name of the bucket to create
        in: path
        name: bucket
        required: true
        type: string
      - description: AWS4-HMAC-SHA256 authorization header (required if S3_AUTH_ENABLED=true)
        in: header
        name: Authorization
        type: string
      - description: Request timestamp in ISO 8601 format (required if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Date
        type: string
      - description: SHA256 hash of request payload or 'UNSIGNED-PAYLOAD' (required
          if S3_AUTH_ENABLED=true)
        in: header
        name: X-Amz-Content-SHA256
        type: string
      produces:
      - text/xml
      responses:
        "200":
          description: Bucket created successfully
          headers:
            Location:
              description: The path to the created bucket
              type: string
        "400":
          description: Invalid bucket name
          schema:
            $ref: '#/definitions/model.S3Error'
        "403":
          description: Access denied or signature mismatch
          schema:
            $ref: '#/definitions/model.S3Error'
        "409":
          description: Bucket already exists
          schema:
            $ref: '#/definitions/model.S3Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.S3Error'
      summary: Create a new bucket
      tags:
      - buckets
...
swagger: "2.0"